+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - nodejs: fix wopi actions after restart
 - setting an unavailable language
 - description in the tooltip on the main page
+- nodejs: referenceData
 - option to send directUrl
 
 ## 1.4.0